organized files, created profile window with initial logic and logout function
